⚡ Bolt: optimize hooks and tracking components for performance
- Memoize return objects for useFunnelTracking, useTaskManagement, useUserPreferences, and useCache.
- Stabilize handleToggleTaskStatus in useTaskManagement using latest-ref pattern.
- Move applyTaskStatusUpdate pure helper outside useTaskManagement hook.
- Consolidate and optimize getSessionId utility using crypto.randomUUID.
- Wrap SessionTracker in React.memo to avoid redundant tracking executions.
- Fix react-hooks/refs lint errors by moving ref assignments to useEffect.
- Add unit tests for useFunnelTracking identity stability.
- Add performance journal entry in .jules/bolt.md.
