+++ src/hooks/useTaskManagement.ts
@@ -44,6 +44,83 @@ interface UseTaskManagementReturn {
   collapseAll: () => void;
 }
 
+// Helper function to apply task status update to state
+// Moved outside hook to ensure static identity and follow best practices
+const applyTaskStatusUpdate = (
+  prevData: TasksResponse | null,
+  taskId: string,
+  newStatus: TaskStatus
+): TasksResponse | null => {
+  if (!prevData) return null;
+
+  const dIndex = prevData.deliverables.findIndex((d) =>
+    d.tasks.some((t) => t.id === taskId)
+  );
+  if (dIndex === -1) return prevData;
+
+  const deliverable = prevData.deliverables[dIndex];
+  const taskIndex = deliverable.tasks.findIndex((t) => t.id === taskId);
+  const task = deliverable.tasks[taskIndex];
+  const est = Number(task.estimate) || 0;
+
+  let deltaTasks = 0;
+  let deltaHours = 0;
+
+  if (newStatus === 'completed' && task.status !== 'completed') {
+    deltaTasks = 1;
+    deltaHours = est;
+  } else if (newStatus !== 'completed' && task.status === 'completed') {
+    deltaTasks = -1;
+    deltaHours = -est;
+  }
+
+  if (deltaTasks === 0) return prevData;
+
+  const updatedTasks = [...deliverable.tasks];
+  updatedTasks[taskIndex] = {
+    ...task,
+    status: newStatus,
+    completion_percentage: newStatus === 'completed' ? 100 : 0,
+  };
+
+  const newCompletedCount = deliverable.completedCount + deltaTasks;
+  const newCompletedHours =
+    Math.round((deliverable.completedHours + deltaHours) * 10) / 10;
+
+  const updatedDeliverable = {
+    ...deliverable,
+    tasks: updatedTasks,
+    completedCount: newCompletedCount,
+    completedHours: newCompletedHours,
+    progress: Math.round(
+      updatedTasks.length > 0 ? (newCompletedCount / updatedTasks.length) * 100 : 0
+    ),
+  };
+
+  const updatedDeliverables = [...prevData.deliverables];
+  updatedDeliverables[dIndex] = updatedDeliverable;
+
+  const { summary } = prevData;
+  const newOverallCompletedTasks = summary.completedTasks + deltaTasks;
+  const newOverallCompletedHours =
+    Math.round((summary.completedHours + deltaHours) * 10) / 10;
+
+  return {
+    ...prevData,
+    deliverables: updatedDeliverables,
+    summary: {
+      ...summary,
+      completedTasks: newOverallCompletedTasks,
+      completedHours: newOverallCompletedHours,
+      overallProgress: Math.round(
+        summary.totalTasks > 0
+          ? (newOverallCompletedTasks / summary.totalTasks) * 100
+          : 0
+      ),
+    },
+  };
+};
+
 export function useTaskManagement(ideaId: string): UseTaskManagementReturn {
   const logger = useMemo(() => createLogger('TaskManagement'), []);
   const [loading, setLoading] = useState(true);
@@ -116,86 +193,6 @@ export function useTaskManagement(ideaId: string): UseTaskManagementReturn {
     }
   }, [ideaId, logger]);
 
-  // Helper function to apply task status update to state
-  const applyTaskStatusUpdate = useCallback(
-    (
-      prevData: TasksResponse | null,
-      taskId: string,
-      newStatus: TaskStatus
-    ): TasksResponse | null => {
-      if (!prevData) return null;
-
-      const dIndex = prevData.deliverables.findIndex((d) =>
-        d.tasks.some((t) => t.id === taskId)
-      );
-      if (dIndex === -1) return prevData;
-
-      const deliverable = prevData.deliverables[dIndex];
-      const taskIndex = deliverable.tasks.findIndex((t) => t.id === taskId);
-      const task = deliverable.tasks[taskIndex];
-      const est = Number(task.estimate) || 0;
-
-      let deltaTasks = 0;
-      let deltaHours = 0;
-
-      if (newStatus === 'completed' && task.status !== 'completed') {
-        deltaTasks = 1;
-        deltaHours = est;
-      } else if (newStatus !== 'completed' && task.status === 'completed') {
-        deltaTasks = -1;
-        deltaHours = -est;
-      }
-
-      if (deltaTasks === 0) return prevData;
-
-      const updatedTasks = [...deliverable.tasks];
-      updatedTasks[taskIndex] = {
-        ...task,
-        status: newStatus,
-        completion_percentage: newStatus === 'completed' ? 100 : 0,
-      };
-
-      const newCompletedCount = deliverable.completedCount + deltaTasks;
-      const newCompletedHours =
-        Math.round((deliverable.completedHours + deltaHours) * 10) / 10;
-
-      const updatedDeliverable = {
-        ...deliverable,
-        tasks: updatedTasks,
-        completedCount: newCompletedCount,
-        completedHours: newCompletedHours,
-        progress: Math.round(
-          updatedTasks.length > 0
-            ? (newCompletedCount / updatedTasks.length) * 100
-            : 0
-        ),
-      };
-
-      const updatedDeliverables = [...prevData.deliverables];
-      updatedDeliverables[dIndex] = updatedDeliverable;
-
-      const { summary } = prevData;
-      const newOverallCompletedTasks = summary.completedTasks + deltaTasks;
-      const newOverallCompletedHours =
-        Math.round((summary.completedHours + deltaHours) * 10) / 10;
-
-      return {
-        ...prevData,
-        deliverables: updatedDeliverables,
-        summary: {
-          ...summary,
-          completedTasks: newOverallCompletedTasks,
-          completedHours: newOverallCompletedHours,
-          overallProgress: Math.round(
-            summary.totalTasks > 0
-              ? (newOverallCompletedTasks / summary.totalTasks) * 100
-              : 0
-          ),
-        },
-      };
-    },
-    []
-  );
 
   // Toggle task status with OPTIMISTIC updates
   const handleToggleTaskStatus = useCallback(
@@ -295,7 +292,7 @@ export function useTaskManagement(ideaId: string): UseTaskManagementReturn {
         setUpdatingTaskId(null);
       }
     },
-    [logger, applyTaskStatusUpdate]
+    [logger]
   );
 
   // Toggle deliverable expansion
@@ -325,15 +322,28 @@ export function useTaskManagement(ideaId: string): UseTaskManagementReturn {
     setExpandedDeliverables(new Set());
   }, []);
 
-  return {
-    loading,
-    error,
-    data,
-    updatingTaskId,
-    expandedDeliverables,
-    handleToggleTaskStatus,
-    toggleDeliverable,
-    expandAll,
-    collapseAll,
-  };
+  return useMemo(
+    () => ({
+      loading,
+      error,
+      data,
+      updatingTaskId,
+      expandedDeliverables,
+      handleToggleTaskStatus,
+      toggleDeliverable,
+      expandAll,
+      collapseAll,
+    }),
+    [
+      loading,
+      error,
+      data,
+      updatingTaskId,
+      expandedDeliverables,
+      handleToggleTaskStatus,
+      toggleDeliverable,
+      expandAll,
+      collapseAll,
+    ]
+  );
 }
